Use ParameterTrait
